Online and offline classifiers 'working', but for short time
Both online and offline detection algorithms are 'working' in this version, in the sense that they get the same results (and AUCs are somewhere around 0.70). However, performance of classifiers drops really quickly in time. Next attempt is to add a whole new section to the software in order to implement Bin He's results (Noninvasive Electroencephalogram Based Control of a Robotic Arm for Reach and Grasp Tasks). This implies no classifier, but just a linear mapping of power to speed in different directions.